+++ CHANGELOG.md
@@ -14,6 +14,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changes
 
+* Improve error log on Draw.io Desktop call
 * Support Draw.io Desktop v14.5.1
   * --embed-diagram is now supported for PDF format
 * Updates default linux path used by Draw.io Desktop since v14.5.1